+++ src/app/_layouts/index.ts
@@ -1,2 +1,2 @@
 export { MainLayout } from "./MainLayout";
-export { Authenticated } from "./Authenticated";
+export { ServerAuthenticated } from "./ServerAuthenticated";
